추천 기능 ì 완성

+++ src/main/java/com/example/demo/Recommendation/QuestionRecommendationKey.java
@@ -12,6 +12,6 @@ import java.io.Serializable;
 @AllArgsConstructor
 @NoArgsConstructor
 public class QuestionRecommendationKey implements Serializable {
-    private Integer questionId;
-    private Integer userId;
+    private Long questionId;
+    private Long userId;
 }
